+++ src/test/java/wooteco/subway/line/dao/LineDaoTest.java
@@ -2,6 +2,7 @@ package wooteco.subway.line.dao;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import static org.junit.jupiter.api.Assertions.*;
 
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
@@ -15,21 +16,21 @@ import wooteco.subway.line.Line;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @Sql("classpath:initializeTable.sql")
-class LineDaoH2Test {
+class LineDaoTest {
 
     @Autowired
-    private LineDaoH2 lineDao;
+    private LineDao lineDao;
 
     @DisplayName("노선 저장")
     @Test
     public void save() {
-        //given
+        // given
         Line line = new Line("10호선", "붉은색");
 
-        //when
-        Line requestedLine = lineDao.save(line);
+        // when
+        Line requestedLine = lineDao.create(line);
 
-        //then
+        // then
         assertThat(requestedLine.getName()).isEqualTo(line.getName());
         assertThat(requestedLine.getColor()).isEqualTo(line.getColor());
     }
@@ -37,60 +38,29 @@ class LineDaoH2Test {
     @DisplayName("노선 중복 저장 시도")
     @Test
     public void duplicatedSave() {
-        //given
+        // given
         Line line1 = new Line("1호선", "초록색");
         Line line2 = new Line("1호선", "파란색");
 
         // when
-        lineDao.save(line1);
-
-        //then
-        assertThatThrownBy(() -> lineDao.save(line2)).isInstanceOf(DuplicateKeyException.class);
-    }
-
-    @Test
-    @DisplayName("주어진 이름과 같은 노선 세기")
-    public void countByName() {
-        String name = "1호선";
-        String color = "파란색";
-        Line line1 = new Line(name, color);
-        lineDao.save(line1);
-
-        // when
-        int number = lineDao.countByName(name);
+        lineDao.create(line1);
 
         // then
-        assertThat(number).isEqualTo(1);
+        assertThatThrownBy(() -> lineDao.create(line2)).isInstanceOf(DuplicateKeyException.class);
     }
-
-    @Test
-    @DisplayName("주어진 색깔과 같은 노선 세기")
-    public void countByColor() {
-        String name = "1호선";
-        String color = "파란색";
-        Line line1 = new Line(name, color);
-        lineDao.save(line1);
-
-        // when
-        int number = lineDao.countByColor(color);
-
-        // then
-        assertThat(number).isEqualTo(1);
-    }
-
 
     @DisplayName("id값에 맞는 노선 반환")
     @Test
     public void findLine() {
-        //given
+        // given
         Line line1 = new Line("12호선", "분홍색");
-        Line saveLine = lineDao.save(line1);
+        Line saveLine = lineDao.create(line1);
         long id = saveLine.getId();
 
-        //when
-        Line requestedLine = lineDao.show(id).get();
+        // when
+        Line requestedLine = lineDao.show(id);
 
-        //then
+        // then
         assertThat(requestedLine.getId()).isEqualTo(id);
         assertThat(requestedLine.getName()).isEqualTo(line1.getName());
         assertThat(requestedLine.getColor()).isEqualTo(line1.getColor());
@@ -98,30 +68,30 @@ class LineDaoH2Test {
 
     @DisplayName("존재하지 않는 id 값을 가진 노선 반환 시도")
     @Test
-    void findLineVoidId() {
-        //given
+    void findLineNotFoundException() {
+        // given
 
-        //when
+        // when
         long id = -1;
 
-        //then
+        // then
         assertThatThrownBy(() -> lineDao.show(id))
             .isInstanceOf(EmptyResultDataAccessException.class);
     }
 
-    @DisplayName("전 노선 호출")
+    @DisplayName("모든 노선 호출")
     @Test
     void findAll() {
-        //given
+        // given
         Line line1 = new Line("10호선", "붉은색");
-        lineDao.save(line1);
+        lineDao.create(line1);
         Line line2 = new Line("11호선", "노란색");
-        lineDao.save(line2);
+        lineDao.create(line2);
 
-        //when
+        // when
         List<Line> lines = lineDao.showAll();
 
-        //then
+        // then
         assertThat(lines.get(0).getName()).isEqualTo(line1.getName());
         assertThat(lines.get(0).getColor()).isEqualTo(line1.getColor());
         assertThat(lines.get(1).getName()).isEqualTo(line2.getName());
@@ -131,19 +101,19 @@ class LineDaoH2Test {
     @DisplayName("노선 업데이트")
     @Test
     void update() {
-        //given
+        // given
         Line line1 = new Line("11호선", "보라색");
-        Line saveLine = lineDao.save(line1);
+        Line saveLine = lineDao.create(line1);
         long id = saveLine.getId();
         String requestName = "분당선";
         String requestColor = "노란색";
         Line requestLine = new Line(requestName, requestColor);
 
-        //when
+        // when
         lineDao.update((int) id, requestLine);
-        Line responseLine = lineDao.show(id).get();
+        Line responseLine = lineDao.show(id);
 
-        //then
+        // then
         assertThat(responseLine.getName()).isEqualTo(requestName);
         assertThat(responseLine.getColor()).isEqualTo(requestColor);
     }
@@ -151,28 +121,28 @@ class LineDaoH2Test {
     @DisplayName("노선 삭제")
     @Test
     void remove() {
-        //given
+        // given
         Line line1 = new Line("12호선", "분홍색");
-        Line saveLine = lineDao.save(line1);
+        Line saveLine = lineDao.create(line1);
         long id = saveLine.getId();
 
-        //when
+        // when
         int number = lineDao.delete(id);
 
-        //then
+        // then
         assertThat(number).isEqualTo(1);
     }
 
-    @DisplayName("노선 삭제 실패")
+    @DisplayName("존재하지 않는 id를 통한 노선 삭제 시도시 예외처리")
     @Test
-    void removeFail() {
-        //given
+    void removeNotFoundException() {
+        // given
         long id = -1;
 
-        //when
+        // when
         int number = lineDao.delete(id);
 
-        //then
+        // then
         assertThat(number).isEqualTo(0);
     }
 }
